QPID-7118: [Java Broker, WMC] Display confirmation on attempt to close modified dashboard/query tabs

+++ broker-plugins/management-http/src/main/java/resources/js/qpid/management/DashboardTab.js
@@ -26,6 +26,7 @@ define(["dojo/parser",
         "qpid/common/util",
         "dojo/text!showDashboardTab.html",
         "qpid/management/dashboard/DashboardWidget",
+        "qpid/common/MessageDialog",
         "dojo/domReady!"],
     function (parser, lang, all, Deferred, query, json, util, template)
     {
@@ -36,16 +37,26 @@ define(["dojo/parser",
             this.tabData = kwArgs.tabData;
             this.parent = kwArgs.tabData.modelObject;
             this.management = this.controller.management;
+            this.changed = !this.tabData.data.name;
+            this.confirmationDilog = new qpid.common.MessageDialog({
+                title: "Discard unsaved changed?",
+                message: "<div>Dashbord contains unsaved changes.<br/>Would you like to close it anyway?</div>"
+            });
+            this.dialogHandle = this.confirmationDilog.on("execute", lang.hitch(this, function (stopDisplaying)
+            {
+                DashboardTab.stopDisplayingConfirmation = stopDisplaying;
+                this.destroy();
+            }));
         }
 
-        DashboardTab.prototype.getTitle = function (changed)
+        DashboardTab.prototype.getTitle = function ()
         {
             if (this.tabData.preferenceId && !this.tabData.data)
             {
                 return "Loading...";
             }
             var name = this.tabData.data.name ? this.tabData.data.name : "New";
-            var prefix = this.tabData.data.name && !changed ? "" : "*";
+            var prefix = this.tabData.data.name && !this.changed ? "" : "*";
             var path = this.controller.structure.getHierarchicalName(this.parent);
             return prefix + "Dashboard:" + name + path;
         };
@@ -116,12 +127,14 @@ define(["dojo/parser",
             this.dashboardWidget.on("save", lang.hitch(this, function(e)
             {
                 this.tabData.data = e.preference;
+                this.changed = false;
                 var title = this.getTitle();
                 this.contentPane.set("title", title);
             }));
             this.dashboardWidget.on("change", lang.hitch(this, function(e)
             {
-                var title = this.getTitle(true);
+                this.changed = true;
+                var title = this.getTitle();
                 this.contentPane.set("title", title);
             }));
             this.dashboardWidget.on("delete", lang.hitch(this, function(e)
@@ -138,25 +151,35 @@ define(["dojo/parser",
                     modelObject: e.parentObject
                 });
             }));
-           // this.dashboardWidget.startup();
+            this.dashboardWidget.startup();
         };
 
         DashboardTab.prototype.close = function ()
         {
-            if (this.dashboardWidget != null)
+            if (!this.changed || (this.changed && DashboardTab.stopDisplayingConfirmation))
             {
-                this.dashboardWidget.destroyRecursive();
-                this.dashboardWidget = null;
+                if (this.dashboardWidget != null)
+                {
+                    this.dashboardWidget.destroyRecursive();
+                    this.dashboardWidget = null;
+                }
+                this.dialogHandle.remove();
+                return true;
             }
+
+            this.confirmationDilog.show();
+            return false;
         };
 
         DashboardTab.prototype.destroy = function ()
         {
-            this.close();
+            this.changed = false;
             this.contentPane.onClose();
             this.controller.tabContainer.removeChild(this.contentPane);
             this.contentPane.destroyRecursive();
         };
+
+        DashboardTab.stopDisplayingConfirmation = false;
 
         return DashboardTab;
     });
